requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
openpyxl==3.1.2
python-dotenv==1.0.0
python-dateutil==2.8.2
